requirements.txt
groq
python-dotenv
requests
numpy